Split Spergel tests into test_intensity_spergel.py
Pure reorganization, no logic changes. Exponential model tests stay in
test_intensity.py (902 lines, 45 tests). Spergel/DeVauc tests move to
test_intensity_spergel.py (1084 lines, 54 tests). Shared fixtures
duplicated in new file.

Also adds spergel_output_dir fixture and inclination diagnostic panel
comparing Spergel vs InclinedSersic at n=1,2,4 x cosi with two mapping
variants (face-on vs inclined).
